requirements.txt
#sentence-transformers==2.2.2
#ctransformers==0.2.5
#pinecone-client
#langchain==0.0.225
#pypdf
langchain
langchain-community
openai
groq
tiktoken
pinecone-client
langchain_pinecone
unstructured
pdfminer==20191125
pdfminer.six==20221105
pillow_heif
unstructured_inference
sentence-transformers
flask

